requirements.txt
flask==2.0.1
google-cloud-pubsub==2.13.0
gunicorn==20.1.0  # WSGI server for Cloud Run